Added basic NavMesh Path Visualization
Added Basic NetField support
Added new command ("cheat path generate")
Fixed ChunkID calculations
Fixed a crash when trying to initialize Spell.cdb
Added premake project lua files to solutions
Updated to new MetaGen System
Updated Submodule Engine

+++ Source/Game-Lib/Game-Lib/ECS/Util/Network/NetworkUtil.h
@@ -35,7 +35,7 @@ namespace ECS
                 ::Network::MessageHeader header =
                 {
                     .opcode = PacketType::PACKET_ID,
-                    .size = packet.GetSerializedSize()
+                    .size = static_cast<u16>(packet.GetSerializedSize())
                 };
 
                 buffer->Put(header);
@@ -59,7 +59,7 @@ namespace ECS
                 ::Network::MessageHeader header =
                 {
                     .opcode = PacketType::PACKET_ID,
-                    .size = packet.GetSerializedSize()
+                    .size = static_cast<u16>(packet.GetSerializedSize())
                 };
 
                 failed |= !buffer->Put(header);
